in remat, handle hoisted outputs in out_avals
In remat's partial_eval rule, we form a TypedJaxpr and hence need to set
its output avals (modeling its output type). Since jaxprs are currently
untyped, the logic looks at the Python output values returned by
call bind to get their types. However, some of the output types might
not correspond to value types in the jaxpr language; for example,
instances of ad_util.Zero can be returned from a call bind. (In the past
we've only formed TypedJaxprs in "initial-style" control-flow
primitives, which don't have this possibility because there's no call
bind involved.)

We handle this possibility just by inspecting the jaxpr first before
looking at the Python-valued outputs: wherever a non-jaxtype value has
been hoisted out, a corresponding unit exists in the jaxpr outvars. So
the logic is now: if the output has an abstract value from partial
evaluation, use that to get the output type; otherwise if the output is
a literal or unitvar, use that to get the output type; otherwise look at
the const from partial evaluation (and it's an error if that const isn't
a valid jaxtype).

fixes #1907