Batching rule for custom_linear_solve (#2099)
* Add batching rule for custom_linear_solve

The custom_linear_solve primitive is batched by batching each of the
jaxprs stored in the primitive. Conceptually, this corresponds to
transforming the initial linear system solve into an implicit "block
diagonal" solve, where matvec/vecmat apply a linear operation to each
part of the input, and solve/transpose_solve solve each block
separately.

Note that the batching of the input and output must agree between all
four jaxprs, since the custom_linear_solve JVP and transpose rules
assume the shapes all match. In particular, the JVP passes the output
of solve into (the JVP of) matvec and then that output back into solve,
and the transpose rule causes the (cotangents for the) output of solve
to be fed back into solve_transpose and vecmat. To ensure consistency
we can do a fixed-point iteration to figure out whether each component
of x and b are batched or not.

* Add support for batched solves without a transpose

If there is no transpose solve, we don't need to batch the transposed
versions of the jaxprs.

* Add pytree test for custom linear solve

Custom linear solve supports solves that act on pytrees, not just single
arrays. This commit adds a test for a linear operator and solver that
operate on Python lists of scalars instead of vectors, and confirms that
transformations work correctly. The batching behavior has been chosen
to make sure it requires a few iterations to find the fixed point of
which elements have a batch axis.