Fix `jax.image._resize` function (#3805)
This PR fixes a bug in jax.image._resize where the local `method_id`
variable may be used without being defined first.
This bug can be easily reproduced by passing to `jax.image.resize`
parameter `method` a `ResizeMethod` instead of an `str`. By doing
this, `method_id` is never defined and the instruction
`if method_id == ResizeMethod.NEAREST` raises an error. Currently,
this can be easily bypassed assigning parameter `method` a `str`.
To fix this bug, it only needs to rename `method_id` to `method`,
the same name of the input parameter.