[jax2tf] Change the tests that assert failure in TF (#4018)
We used to have some tests that asserted for failures in TF after conversion.
The idea was for the tests to be a living record of what is failing. The
problem is that this makes it hard to fix those TF failures, because one
would have to fix at the same time the test.

This change just ensures that we tolerate TF exceptions without actually
ensuring that they arise.