Update auto jaxlib install to use nvcc CUDA version instead of nvidia-smi. (#4050)
The CUDA version reported by `nvidia-smi` is based on the driver
version, not the runtime/toolkit version, and may be higher than the
toolkit version since the driver supports older runtimes. `nvcc`,
which is also installed as part of the CUDA toolkit, will correctly
report the right CUDA version to target.

Fixes #3984.