Add a function to get the signature associated to an argument.
For the C++ jax.jit, we were able to first perform the transfer, and then inspect the result to know what the dtype and shape are.
However, for pmap, we don't necessarily know how to transfer the arguments, before having inspected the signature and retrieved the compiled executable & sharding_sec.

So for pmap, we will do explicitly the "abstractify" and the "shard_args/device_put" steps.

PiperOrigin-RevId: 351461953