add transformed fun src info to escaped tracer err
This change adds to the error message when we hit an escaped tracer. In
particular, it adds source info for the function that was transformed.

This change currently only applies to escaped `DynamicJaxprTracer`s
(arising from `jit`, `pmap`, `scan`, and other staging functions) and
not other traces. A natural follow-up would be to attach this
information to other traces.

Co-authored-by: Lena Martens <lenamartens@google.com>