Replace `host_id` with `process_index` terminology, take 2.
We're switching to the new terminology to avoid confusion in cases
where multiple jax processes are running on a single host, and each
process has a unique process_index/host_id.

This keeps aliases for the old `host_id` APIs for now, but these will
eventually be removed.

This was originally commited in
6044772c7ca7aaaf7c7aa6dcb261db713c35acb5, but reverted in
022774cce0cdda446a491f8b5ff430eddd10b76a due to Google-internal test
failures from renaming the local_devices argument name. This change is
identical except it also adds staging for the argument name change.