[jax2tf] Improve conversion of pad when enable_xla=True, add tests.
If enable_xla we should directly use the XLA conversion, without
trying to see whether the primitive can actually be converted
to non XLA TF ops. This is happening elsewhere, but it was not
happening for pad.

Also discovered that we were never turning enable_xla=False in
tests, even when we had special harnesses for it.