Tighten the numerical tolerances for jax2tf.
We want to focus on tolerances for the TF compiled mode. So, we
make the `custom_numeric` apply by default only to `eager` and `graph` modes.

Getting this done required iterating through tests and we used this
opportunity to tighten the tolerances and apply
them more narrowly, e.g., only on CPU or GPU. The tolerances are still
somewhat loose, mainly for linear algebra primitives.

PiperOrigin-RevId: 374215224