[jax2tf] Add a best-effort conversion for gather with enable_xla=False
In an attempt to support lax.gather in absence of XLA, e.g., for
TFjs converter, we add an alternative conversion for the case
when enable_xla=False. Previously, the conversion would fail in this case.

The important parts of the code are lifeted from an old PR #3486, which was
then removed in #4030 (to avoid having to think about out-of-bounds accesses).

There is still a discrepancy between the JAX and TF behavior for out-of-bounds
accesses.