Add `pip install jax[tpu]` configuration.
This can be used on Cloud TPU VMs to automatically install compatible
versions of jax, jaxlib, and libtpu (the low-level library JAX uses to
access the TPU on Cloud TPU VMs).

The new install command requires a new jax release (`>=0.2.15`) and
jaxlib release (`>=0.1.68`) to work, since it requires both
https://github.com/google/jax/commit/15f8232136baabeee56f25a7976575e396abf54d
and
https://github.com/tensorflow/tensorflow/commit/ce2bc249968344c57e91c45799ad5ec91aad2566
to pick up the pip-installed libtpu. I'll update the README and Cloud
TPU VM documentation once these releases are out.