[call_tf] Improved call_tf for op-by-op executions.
There are two major improvements here. First we ensure that
in op-by-op execution we can even execute functions that are not
compileable. We do this by ensuring that we do not trace the
TF function to a graph too early.

The other improvement is to work around some bugs in the TF shape
inference. Some TF graphs has unknown output shapes even when traced
with known inputs shapes. This happens even for some graph that
are generated by jax2tf, which we know should have known shapes. To
work around this, we get the output shapes for the TF function using
the XLA compiler, which is more reliably able to figure out the output
shapes. We do this even during abstract evaluation of the call_tf
primitive, and we use caching to ensure we do not call the TF
compiler repeatedly.