Add `jax_hlo_source_file_canonicalization_regex` config.
This is meant to be used with @colemanliyah's persistent compilation
cache, since the serialized HLO computation (including the source_file
metadata) is used in the cache key. The config can be used to remove
bits of the source file path that vary between program invocations, to
avoid spurious cache misses.